% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Matrix Multiplication using RcppEigen}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{numeric matrix of dimension \eqn{m \times n}}

\item{B}{numeric matrix of dimension \eqn{n \times l}}
}
\value{
C matrix of dimension \eqn{m \times p} of the row-column product of A and B and C
}
\description{
Matrix multiplication of the two matrices in input, without copy.
}
