% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_distance_matrix.R
\name{get_mean_distance_matrix}
\alias{get_mean_distance_matrix}
\alias{getMeanDistanceMatrix}
\title{Mean distance matrix}
\usage{
get_mean_distance_matrix(DM_list)

getMeanDistanceMatrix(DM_list)
}
\arguments{
\item{DM_list}{list of distance matrices}
}
\value{
mean-distance matrix
}
\description{
Given a sequence of distance matrices, expected to correspond to sequential increasing values of time, calculate the average distance between any pairs of nodes
shortest-path (or geodesic) distance
}
\section{Functions}{
\itemize{
\item \code{getMeanDistanceMatrix()}: Old deprecated function

}}
\references{
De Domenico, M. (2017). Diffusion Geometry Unravels the Emergence of
Functional Clusters in Collective Phenomena. Physical Review Letters.
\doi{10.1103/PhysRevLett.118.168301}

Bertagnolli, G., & De Domenico, M. (2020). Diffusion Geometry of Multiplex
and Interdependent Systems.
\href{https://arxiv.org/abs/2006.13032}{arxiv preprint arxiv:2006.13032}
}
\seealso{
\code{\link{get_diffusion_probability_matrix}}, \code{\link{get_distance_matrix}}
}
\keyword{distance}
