% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_diffudist.R
\name{plot_distance_matrix}
\alias{plot_distance_matrix}
\title{Plot distance matrix}
\usage{
plot_distance_matrix(
  DM,
  col_palette = viridis::viridis(n = 11),
  log_scale = F,
  cex = 1,
  show_dendro = TRUE,
  title = ""
)
}
\arguments{
\item{DM}{a distance matrix}

\item{col_palette}{a colour palette, previously it was set to the `spectral`
palette of [RColorBrewer::brewer.pal] but now it is, by default, to the
color-blind firendly [viridis::viridis].}

\item{log_scale}{logical. Default FALSE}

\item{cex}{numerical value by which text should be magnified (default font
size in [ggplot2:theme] is 11)}

\item{show_dendro}{If the dendrogram resulting from [stats::hclust] should
be shown. Default TRUE}

\item{title}{Title of the plot passed to [ggplot]. No title by default.}
}
\value{
plot [ggplot]
}
\description{
Plot a heatmap of the distance matrix using ggplot. Useful if used with multiplot ([source](http://stackoverflow.com/questions/6673162/reproducing-lattice-dendrogram-graph-with-ggplot2))
}
\keyword{distance}
\keyword{heatmap}
\keyword{matrix;}
\keyword{plot;}
