% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_diffudist.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot distance matrix as heatmap}
\usage{
plotHeatmap(
  DM,
  colPalette = NULL,
  log.scale = FALSE,
  cex = 1,
  showDendrogram = TRUE,
  title = ""
)
}
\arguments{
\item{DM}{a distance matrix}

\item{colPalette}{`r lifecycle::badge("deprecated")` in the new function \link{plot_distance_matrix}
it is called \code{col_palette}}

\item{log.scale}{`r lifecycle::badge("deprecated")` in the new function \link{plot_distance_matrix}
it is called \code{log_scale}}

\item{cex}{numerical value by which text should be magnified (default font
size in [ggplot2:theme] is 11)}

\item{showDendrogram}{`r lifecycle::badge("deprecated")` in the new function \link{plot_distance_matrix}
it is called \code{show_dendro}}

\item{title}{Title of the plot passed to \link[ggplot2]{labs}. No title by
default}
}
\value{
a \link[ggplot2]{ggplot}
}
\description{
`r lifecycle::badge("deprecated")`
}
